Benchmark result summary created
